requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
langchain
python-dotenv
openai
huggingface_hub
pytest
requests
